+++ src/main/java/com/muzzy/domain/AncestorTransaction.java
@@ -4,6 +4,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.security.PublicKey;
 import java.util.ArrayList;
 
@@ -12,7 +13,7 @@ import java.util.ArrayList;
 public class AncestorTransaction extends Transaction {
 
     @Builder(builderMethodName = "childBuilder")
-    public AncestorTransaction(PublicKey sender, PublicKey reciever, float value, ArrayList<TransactionInput> inputs,ArrayList<TransactionOutput> outputs, String transactionId) {
+    public AncestorTransaction(PublicKey sender, PublicKey reciever, BigDecimal value, ArrayList<TransactionInput> inputs, ArrayList<TransactionOutput> outputs, String transactionId) {
         super(sender, reciever, value, inputs, outputs, transactionId);
     }
